py2many: infer the type of array elements
    a: List[int] = [1, 2, 3]
    b = a[1]

b is inferred to be an int

+++ tests/expected/binit.rs
@@ -11,7 +11,7 @@ pub fn bisect_right(data: &Vec<i32>, item: i32) -> i32 {
     let mut high: i32 = data.len() as i32;
     while low < high {
         let middle: _ = i32::from(((low + high) / 2));
-        if item < data[middle as usize] as i32 {
+        if item < data[middle as usize] {
             high = middle;
         } else {
             low = (middle as i32 + 1);
